requirement.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
django-jazzmin==2.6.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.4
psycopg2-binary==2.9.7
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
Pillow==10.2.0